ASCA

This function performs the full ASCA decomposition and acts as a constructor of the ASCA S3 object

ASCA(d, x, f, comb = NULL, glm_par = vector(mode = "list", length = 0))

# S3 method for asca
print(x)

# S3 method for asca
summary(x)

Arguments

d
a data frame with the experimental design, which associate each sample to the design factors

x
a matrix with the data measured on the different samples

f
a string with the formula which will be used for the glm decomposition

comb
a character vector specifying which terms of the decomposition should be combined

glm_par
a list with the parameters to be passed to the glm call

Value

An object of class asca. Actually a list holding the results:

of the glm decomposition

of the svd of the factor matrices

of the validation of the object (if present)

Details

The ASCA decomposition of a data matrix is performed by using Generalized Linear Models to perform the estimation of the expected values. the use of GLM's allows the extension of the method to non normal data and unbalanced designs

Note

There are specific methods to handle and extract information from the asca object

print

summary

scores

loadings

varimp

plot.asca

validate

Examples

## load the data
data("synth_count_data")

## perform the ASCA decomposition
ASCA_test <- ASCA(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
glm_par = list(family = poisson())
)

## decomposition with a combination of treatment and interaction
ASCA_test_comb <- ASCA(
d = synth_count_data$design,
x = synth_count_data$counts, 
f = "time + treatment + time:treatment",
comb = c("treatment","time:treatment"),
glm_par = list(family = poisson())
)

## Print a summary of the decomposition
print(ASCA_test)

#> ASCA decomposition 
#> 
#> Number of samples  45 
#> Number of variables 10 
#> 
#> Formula:  time + treatment + time:treatment 
#> Terms: time,treatment,time:treatment,error

summary(ASCA_test)

#> ASCA decomposition 
#> 
#> Number of samples  45 
#> Number of variables 10 
#> 
#> Formula:  time + treatment + time:treatment 
#> Terms: time,treatment,time:treatment,error
#> Validation:  FALSE 
#> 
#> glm parameters: 
#> $family
#> 
#> Family: poisson 
#> Link function: log 
#> 
#>